Добавить отдельную операцию изменение пароля и рефакторинг UserService

+++ src/main/java/kg/musabaev/em_bank_rest/service/impl/SimpleUserService.java
@@ -2,13 +2,16 @@ package kg.musabaev.em_bank_rest.service.impl;
 
 import kg.musabaev.em_bank_rest.dto.GetCreatePatchUserResponse;
 import kg.musabaev.em_bank_rest.dto.PatchUserRequest;
+import kg.musabaev.em_bank_rest.dto.UpdatePasswordRequest;
 import kg.musabaev.em_bank_rest.entity.User;
+import kg.musabaev.em_bank_rest.exception.PasswordValidationException;
 import kg.musabaev.em_bank_rest.exception.UserAlreadyExistsException;
 import kg.musabaev.em_bank_rest.exception.UserNotFoundException;
 import kg.musabaev.em_bank_rest.mapper.UserMapper;
 import kg.musabaev.em_bank_rest.repository.UserRepository;
 import kg.musabaev.em_bank_rest.repository.specification.UserSpecification;
 import kg.musabaev.em_bank_rest.security.SimpleUserDetails;
+import kg.musabaev.em_bank_rest.service.AuthService;
 import kg.musabaev.em_bank_rest.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
@@ -24,6 +27,7 @@ public class SimpleUserService implements UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
     private final PasswordEncoder passwordEncoder;
+    private final AuthService authService;
 
     @Override
     @Transactional(readOnly = true)
@@ -34,50 +38,65 @@ public class SimpleUserService implements UserService {
 
     @Override
     @Transactional
-    public GetCreatePatchUserResponse patch(Long id, PatchUserRequest dto) {
+    public GetCreatePatchUserResponse patchForAdmin(Long id, PatchUserRequest dto) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException(id));
-        if (user.getEmail().equals(dto.email())) {
-            return patchUser(dto, user);
-        }
-        else if (userRepository.existsByEmail(dto.email()))
+
+        if (!user.getEmail().equalsIgnoreCase(dto.email())
+                && userRepository.existsByEmail(dto.email()))
             throw new UserAlreadyExistsException();
-        else
-            return patchUser(dto, user);
+
+        return patchUser(dto, user);
     }
 
     private GetCreatePatchUserResponse patchUser(PatchUserRequest dto, User user) {
         userMapper.patch(dto, user);
-        user.setPassword(passwordEncoder.encode(dto.password()));
         var persistedUser = userRepository.saveAndFlush(user);
         return userMapper.toPatchUserResponse(persistedUser);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public GetCreatePatchUserResponse getById(SimpleUserDetails userDetails) {
+    public GetCreatePatchUserResponse getByIdForUser(SimpleUserDetails userDetails) {
         var authUser = userDetails.getUser();
         return userMapper.toGetUserResponse(authUser);
     }
 
     @Override
     @Transactional
-    public GetCreatePatchUserResponse patch(PatchUserRequest dto, SimpleUserDetails userDetails) {
+    public GetCreatePatchUserResponse patchForUser(PatchUserRequest dto, SimpleUserDetails userDetails) {
         var authUser = userDetails.getUser();
         return patchUser(dto, authUser);
     }
 
     @Override
     @Transactional
+    public void updatePassword(UpdatePasswordRequest dto, SimpleUserDetails userDetails) {
+        var authUser = userDetails.getUser();
+        var oldPassword = dto.oldPassword();
+        var newPassword = dto.newPassword();
+
+        if (!passwordEncoder.matches(oldPassword, authUser.getPassword()))
+            throw new PasswordValidationException("Invalid old password.");
+
+        authUser.setPassword(passwordEncoder.encode(newPassword));
+
+        userRepository.save(authUser);
+
+        authService.revokeAllUserRefreshTokens(authUser.getId());
+    }
+
+    @Override
+    @Transactional
     public void delete(Long id) {
-        userRepository.findById(id)
-                .orElseThrow(() -> new UserNotFoundException(id));
+        if (userRepository.existsById(id))
+                throw new UserNotFoundException(id);
         userRepository.deleteById(id);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public GetCreatePatchUserResponse getById(Long id) {
+    public GetCreatePatchUserResponse getByIdForAdmin(Long id) {
         return userMapper.toGetUserResponse(userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException(id)));
     }
